model/requirements.txt
decorator==4.4.2
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.5
opencv-python==4.5.2.52
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.13.1
scipy==1.5.4
six==1.16.0
torch==1.8.1
torchvision==0.9.1

